requirements.txt
altgraph==0.16.1
asn1crypto==1.0.1
bcrypt==3.1.7
cffi==1.12.3
Click==7.0
cryptography==2.7
future==0.18.1
paramiko==2.6.0
pefile==2019.4.18
Pillow==6.2.1
pycparser==2.19
PyInstaller==3.5
PyMySQL==0.9.3
PyNaCl==1.3.0
PyQt5==5.11.3
PyQt5-sip==4.19.17
pyqt5-tools==5.11.3.1.4
python-dotenv==0.10.3
pywin32==225
pywin32-ctypes==0.2.0
selenium==3.141.0
sip==4.19.8
six==1.12.0
sshtunnel==0.1.5
urllib3==1.25.6
